Move tour pydantic models to their own module
